pyproject.toml
[project]
requires-python = ">=3.12"
name = "filelink"
version = "0.1.0"
dependencies = ["Django", "django-sendfile2"]

[project.optional-dependencies]
dev = ["autopep8", "pip-tools", "pytest-django"]

[tool.setuptools]
py-modules=[]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "filelink.settings.dev"
